requirements.txt
absl-py==0.9.0
astunparse==1.6.3
bleach==1.5.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
enum34==1.1.10
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
html5lib==0.9999999
idna==2.9
importlib-metadata==1.6.1
Keras-Preprocessing==1.1.2
keyboard==0.13.5
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyQt5==5.13.0
PyQt5-sip==12.8.0
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.13.0.1.5
python-dotenv==0.13.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.2.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.4.0
tensorflow-estimator==2.2.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
